+++ R/prepare_results_speMCA.R
@@ -11,13 +11,22 @@ prepare_results.speMCA <- function(obj) {
 
     if (!inherits(obj, "speMCA")) stop("obj must be of class speMCA")
 
+    ## Extract variable names from results row names
+    extract_var <- function(df) {
+        gsub("(.*)____.*?$", "\\1", rownames(df))
+    }
+    ## Extract level names from results row names
+    extract_mod <- function(df) {
+        gsub(".*____(.*?)$", "\\1", rownames(df))
+    }
+
     vars <- data.frame(obj$var$coord)
     ## Axes names and inertia
     axes <- seq_len(ncol(obj$var$coord))
     names(axes) <- paste("Axis", axes, paste0("(", head(round(obj$eig$rate, 2), length(axes)),"%)"))
     ## Eigenvalues
     eig <- data.frame(dim = 1:length(obj$eig$rate), percent = obj$eig$rate)
-    
+
     ## Variables coordinates
     varnames <- sapply(obj$call$X[,obj$call$quali, drop = FALSE], nlevels)
     varnames <- rep(names(varnames),varnames)
@@ -26,26 +35,42 @@ prepare_results.speMCA <- function(obj) {
     vars$modname <- rownames(vars)
     vars$Type <- "Active"
     vars$Class <- "Qualitative"
+    if (!is.null(obj$supv)) {
+        tmp_sup <- data.frame(obj$supv$coord)
+        tmp_sup$varname <- extract_var(tmp_sup)
+        tmp_sup$modname <- extract_mod(tmp_sup)
+        tmp_sup$Type <- "Supplementary"
+        tmp_sup$Class <- "Qualitative"
+        vars <- vars %>% bind_rows(tmp_sup)
+    }
 
     vars <- vars %>% pivot_longer(names_to = "Axis", values_to = "Coord", starts_with("dim.")) %>%
         mutate(Axis = gsub("dim.", "", Axis, fixed = TRUE),
                Coord = round(Coord, 3))
 
-    ## Contributions
+    ## Variables contrib
     tmp <- data.frame(obj$var$contrib)
     tmp <- tmp %>% mutate(modname = rownames(tmp), Type = "Active", Class = "Qualitative") %>%
         pivot_longer(names_to = "Axis", values_to = "Contrib", starts_with("dim.")) %>%
         mutate(Axis = gsub("dim.", "", Axis, fixed = TRUE),
                Contrib = round(Contrib, 3))
-    
+
     vars <- vars %>% left_join(tmp, by = c("modname", "Type", "Class", "Axis"))
-    
-    ## Cos2
+
+    ## Variables cos2
     tmp <- data.frame(obj$var$cos2)
     tmp$modname <- rownames(tmp)
     tmp$Type <- "Active"
     tmp$Class <- "Qualitative"
-    tmp <- tmp %>% pivot_longer(names_to = "Axis", values_to = "Cos2", starts_with("dim.")) %>%
+    if (!is.null(obj$supv)) {
+        tmp_sup <- data.frame(obj$supv$cos2)
+        tmp_sup$modname <- extract_mod(tmp_sup)
+        tmp_sup$Type <- "Supplementary"
+        tmp_sup$Class <- "Qualitative"
+        tmp <- tmp %>% bind_rows(tmp_sup)
+    }
+    tmp <- tmp %>%
+        pivot_longer(names_to = "Axis", values_to = "Cos2", starts_with("dim.")) %>%
         mutate(Axis = gsub("dim.", "", Axis, fixed = TRUE),
                Cos2 = round(Cos2, 3))
     vars <- vars %>% left_join(tmp, by = c("modname", "Type", "Class", "Axis"))
@@ -57,13 +82,14 @@ prepare_results.speMCA <- function(obj) {
         rename(Variable = varname, Level = modname) %>%
         mutate(Count = NA) %>%
         as.data.frame()
-    
+
     ## Variables eta2
     vareta2 <- data.frame(obj$var$eta2)
     vareta2$Variable <- rownames(vareta2)
     vareta2$Type <- "Active"
     vareta2$Class <- "Qualitative"
-    vareta2 <- vareta2 %>% pivot_longer(names_to = "Axis", values_to = "eta2", starts_with("dim.")) %>%
+    vareta2 <- vareta2 %>%
+        pivot_longer(names_to = "Axis", values_to = "eta2", starts_with("dim.")) %>%
         mutate(Axis = gsub("dim.", "", Axis, fixed = TRUE))
     vareta2$eta2 <- format(vareta2$eta2, scientific = FALSE, nsmall = 3, digits = 0)
 
@@ -77,7 +103,8 @@ prepare_results.speMCA <- function(obj) {
         tmp_sup$Type <- "Supplementary"
         ind <- ind %>% bind_rows(tmp_sup)
     }
-    ind <- ind %>% pivot_longer(names_to = "Axis", values_to = "Coord", starts_with("dim.")) %>%
+    ind <- ind %>%
+        pivot_longer(names_to = "Axis", values_to = "Coord", starts_with("dim.")) %>%
         mutate(Axis = gsub("dim.", "", Axis, fixed = TRUE),
                Coord = round(Coord, 3))
 
@@ -87,13 +114,14 @@ prepare_results.speMCA <- function(obj) {
         pivot_longer(names_to = "Axis", values_to = "Contrib", starts_with("dim.")) %>%
         mutate(Axis = gsub("dim.", "", Axis, fixed = TRUE),
                Contrib = round(Contrib, 3))
-    
+
     ind <- ind %>% left_join(tmp, by = c("Name", "Type", "Axis"))
 
     ## Individuals Cos2
     if (!is.null(obj$supi)) {
         tmp <- data.frame(obj$supi$cos2)
-        tmp <- tmp %>% mutate(Name = rownames(tmp), Type = "Supplementary") %>%
+        tmp <- tmp %>%
+            mutate(Name = rownames(tmp), Type = "Supplementary") %>%
             pivot_longer(names_to = "Axis", values_to = "Cos2", starts_with("dim.")) %>%
             mutate(Axis = gsub("dim.", "", Axis, fixed = TRUE),
                    Cos2 = round(Cos2, 3))
@@ -101,15 +129,18 @@ prepare_results.speMCA <- function(obj) {
     } else {
         ind$Cos2 <- NA
     }
-    
+
     ## Qualitative data for individuals plot color mapping
     quali_data <- obj$call$X[,obj$call$quali]
     if (!is.null(obj$quali.sup)) {
-        quali_data <- quali_data %>% bind_cols(obj$call$X[,obj$call$quali.sup, drop = FALSE])
+        quali_data <- quali_data %>%
+            bind_cols(obj$call$X[,obj$call$quali.sup, drop = FALSE])
     }
     quali_data$Name <- rownames(obj$call$X)
-    
-    
-    return(list(vars = vars, ind = ind, eig = eig, axes = axes, vareta2 = vareta2, quali_data = quali_data))
-    
+
+
+    return(
+        list(vars = vars, ind = ind, eig = eig, axes = axes, vareta2 = vareta2, quali_data = quali_data)
+    )
+
 }